+++ sbercloud/resource_sbercloud_identity_group_membership_test.go
@@ -9,7 +9,7 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/terraform"
 
 	"github.com/huaweicloud/golangsdk/openstack/identity/v3/users"
-	"github.com/huaweicloud/terraform-provider-huaweicloud/huaweicloud"
+	"github.com/huaweicloud/terraform-provider-huaweicloud/huaweicloud/config"
 )
 
 func TestAccIdentityV3GroupMembership_basic(t *testing.T) {
@@ -48,7 +48,7 @@ func TestAccIdentityV3GroupMembership_basic(t *testing.T) {
 }
 
 func testAccCheckIdentityV3GroupMembershipDestroy(s *terraform.State) error {
-	config := testAccProvider.Meta().(*huaweicloud.Config)
+	config := testAccProvider.Meta().(*config.Config)
 	identityClient, err := config.IdentityV3Client(SBC_REGION_NAME)
 	if err != nil {
 		return fmt.Errorf("Error creating SberCloud identity client: %s", err)
@@ -80,7 +80,7 @@ func testAccCheckIdentityV3GroupMembershipExists(n string, us []string) resource
 			return fmt.Errorf("No ID is set")
 		}
 
-		config := testAccProvider.Meta().(*huaweicloud.Config)
+		config := testAccProvider.Meta().(*config.Config)
 		identityClient, err := config.IdentityV3Client(SBC_REGION_NAME)
 		if err != nil {
 			return fmt.Errorf("Error creating SberCloud identity client: %s", err)