+++ sbercloud/resource_sbercloud_vpc_bandwidth_test.go
@@ -9,7 +9,7 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/terraform"
 
 	"github.com/huaweicloud/golangsdk/openstack/networking/v1/bandwidths"
-	"github.com/huaweicloud/terraform-provider-huaweicloud/huaweicloud"
+	"github.com/huaweicloud/terraform-provider-huaweicloud/huaweicloud/config"
 )
 
 func TestAccVpcBandWidthV2_basic(t *testing.T) {
@@ -70,7 +70,7 @@ func TestAccVpcBandWidthV2_WithEpsId(t *testing.T) {
 }
 
 func testAccCheckVpcBandWidthV2Destroy(s *terraform.State) error {
-	config := testAccProvider.Meta().(*huaweicloud.Config)
+	config := testAccProvider.Meta().(*config.Config)
 	networkingClient, err := config.NetworkingV1Client(SBC_REGION_NAME)
 	if err != nil {
 		return fmt.Errorf("Error creating sbercloud networking client: %s", err)
@@ -101,7 +101,7 @@ func testAccCheckVpcBandWidthV2Exists(n string, bandwidth *bandwidths.BandWidth)
 			return fmt.Errorf("No ID is set")
 		}
 
-		config := testAccProvider.Meta().(*huaweicloud.Config)
+		config := testAccProvider.Meta().(*config.Config)
 		networkingClient, err := config.NetworkingV1Client(SBC_REGION_NAME)
 		if err != nil {
 			return fmt.Errorf("Error creating sbercloud networking client: %s", err)